Implement CallSc
