+++ wails-frontend/src/pages/PasswordPromptHandler/Default.tsx
